Change PokeBot to EeveeBot and add a little personality to Eeveebot
