--- src/main/java/duke/ToDos.java
@@ -1,13 +1,0 @@
-package duke;
-
-public class ToDos extends Task {
-
-    public ToDos(String name) {
-        super(name, "T", null, null);
-    }
-
-    public ToDos(String name, boolean isDone) {
-        super(name, "T", null, null, isDone);
-    }
-
-}
