+++ src/main/java/eevee/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package eevee;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -55,14 +55,14 @@ public class DialogBox extends HBox {
     }
 
     public static DialogBox getHelloMessage(Image img) {
-        return getDukeDialog("SUP, I'm PokeBot! How can I help you today?", img);
+        return getEeveeDialog("EEVEE! I'm EeveeBot! How can I help you today?", img);
     }
 
     public static DialogBox getUserDialog(String text, Image img) {
         return new DialogBox(text, img);
     }
 
-    public static DialogBox getDukeDialog(String text, Image img) {
+    public static DialogBox getEeveeDialog(String text, Image img) {
         var db = new DialogBox(text, img);
         db.flip();
         return db;
