+++ src/main/java/eevee/EeveeException.java
@@ -1,0 +1,7 @@
+package eevee;
+
+public class EeveeException extends Exception{
+    public EeveeException(String message) {
+        super(message);
+    }
+}
