+++ src/main/java/eevee/Launcher.java
@@ -1,4 +1,4 @@
-package duke;
+package eevee;
 
 import javafx.application.Application;
