+++ src/main/java/eevee/Main.java
@@ -1,4 +1,4 @@
-package duke;
+package eevee;
 
 import java.io.IOException;
 
@@ -9,11 +9,11 @@ import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
 /**
- * A GUI for Duke using FXML.
+ * A GUI for Eevee using FXML.
  */
 public class Main extends Application {
 
-    private Duke duke = new Duke("./data/tasks.txt");
+    private Eevee eevee = new Eevee("./data/tasks.txt");
 
     @Override
     public void start(Stage stage) {
@@ -22,8 +22,8 @@ public class Main extends Application {
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
             stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            stage.setTitle("PokeBot");
+            fxmlLoader.<MainWindow>getController().setEevee(eevee);
+            stage.setTitle("EeveeBot");
             stage.show();
         } catch (IOException e) {
             e.printStackTrace();
