+++ src/main/java/eevee/ToDos.java
@@ -1,0 +1,13 @@
+package eevee;
+
+public class ToDos extends Task {
+
+    public ToDos(String name) {
+        super(name, "T", null, null);
+    }
+
+    public ToDos(String name, boolean isDone) {
+        super(name, "T", null, null, isDone);
+    }
+
+}
