+++ src/main/resources/view/DialogBox.fxml
@@ -8,14 +8,17 @@
 
 <fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="-Infinity" prefWidth="400.0" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
     <children>
-        <Label fx:id="dialog" text="Label" wrapText="true">
+        <Label fx:id="dialog" style="-fx-border-color: black;" text="Label" wrapText="true">
          <font>
-            <Font name="Constantia" size="11.0" />
+            <Font name="Constantia" size="12.0" />
          </font>
          <padding>
-            <Insets right="5.0" left="5.0" />
-         </padding></Label>
-        <ImageView fx:id="displayPicture" fitHeight="75.0" fitWidth="75.0" pickOnBounds="true" preserveRatio="true" />
+            <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+         </padding>
+         <HBox.margin>
+            <Insets left="5.0" right="5.0" />
+         </HBox.margin></Label>
+        <ImageView fx:id="displayPicture" fitHeight="89.0" fitWidth="89.0" pickOnBounds="true" preserveRatio="true" />
     </children>
     <padding>
         <Insets bottom="10.0" left="5.0" right="5.0" top="10.0" />
