+++ src/test/java/eevee/TaskTest.java
@@ -1,4 +1,4 @@
-package duke;
+package eevee;
 
 import org.junit.jupiter.api.Test;
